fix: throughput only shows real inference activity, no fake spikes
Removed slot-polling based fake throughput. Now uses real tok/s from
llama-server's timings in the relay response. Throughput is only
non-zero while actively processing a relay request. Prevents phantom
spikes that would confuse providers about earnings.

+++ crates/compute-daemon/src/hardware.rs
@@ -340,7 +340,7 @@ fn collect_macos_metrics() -> (Option<u32>, Option<u32>) {
         .and_then(|v| v.trim().parse::<u64>().ok())
         .map(|raw| {
             let current_ma = if raw > (1u64 << 63) {
-                (raw.wrapping_neg()) as u64 // absolute value of negative
+                raw.wrapping_neg() // absolute value of negative
             } else {
                 raw
             };
